feat: 实现 MVP 核心功能并修复全部代码审查问题
## 核心功能实现
- canvas_screen: 实现完整保存流程（序列化笔画 + 截图 + 数据库 + OCR 解析入库）
- canvas_screen: 接入真实 OCR 引擎（MlKit/Vision），替换硬编码模拟数据
- canvas_screen: 实现笔记加载/编辑（从数据库恢复笔画和 OCR 结果）

## Bug 修复
- entry_parser: 修复下下周/下周、大后天/后天匹配顺序
- entry_parser: 补全任务关键词（记得/需要/别忘了/提醒等）+ 本周周X日期匹配
- database_helper: 启用 PRAGMA foreign_keys = ON，version 升至 2 并迁移 amount REAL→TEXT
- image_storage: saveThumbnail 真正缩放至 200×200，修复橡皮擦颜色为 transparent
- note_list_bloc: 删除笔记时先清理图片文件，UUID 替换 timestamp 作为笔记 ID
- search_screen + note_list_screen: 修复搜索防抖（Future.delayed→Timer，接入 NoteListBloc）
- app.dart: 移除多余的全局 CanvasBloc Provider
- entry_row: amount 显示改为 toStringAsFixed(2) 适配 Decimal 类型

## 测试新增
- test/bloc/canvas_bloc_test.dart: 28 个用例（初始状态/笔画/undo-redo/序列化/工具切换）
- test/bloc/note_list_bloc_test.dart: 13 个用例（加载/创建/删除/搜索，内存数据库）
- test/parser/entry_parser_test.dart: 新增 6 个任务关键词和周X日期用例

+++ lib/core/models/note_entry.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:equatable/equatable.dart';
 
 enum NoteEntryType {
@@ -28,7 +29,7 @@ class NoteEntry extends Equatable {
 }
 
 class ExpenseRecord extends Equatable {
-  final double amount;
+  final Decimal amount;
   final String category;
   final String description;
